scripts/requirements.txt
pandas
httpx
python-dotenv
psycopg2-binary
